refactor workflow out of core into cli
