+++ packages/cli/src/lib/index.ts
@@ -10,3 +10,4 @@ export * from "./CacheDirectory";
 export * from "./Compiler";
 export * from "./CodeGenerator";
 export * from "./SchemaComposer";
+export * from "./workflow";
