+++ packages/js/core/src/index.ts
@@ -4,4 +4,3 @@ export * from "./interfaces";
 export * from "./uri-resolution/core";
 export * from "./uri-resolution/resolvers";
 export * from "./utils";
-export * from "./workflow";
