+++ packages/js/core/src/types/index.ts
@@ -11,4 +11,3 @@ export * from "./Env";
 export * from "./InterfaceImplementations";
 export * from "./PluginRegistration";
 export * from "./UriResolver";
-export * from "./Workflow";
